AttributesDTO conversion
